backend/requirements.txt
# WAR ROOM Backend Dependencies
google-adk
google-cloud-firestore
google-cloud-pubsub
google-generativeai
google-genai
fastapi
uvicorn[standard]
websockets
pydantic
pydantic-settings
python-dotenv
livekit-agents[google,elevenlabs]
livekit-plugins-google
livekit-plugins-elevenlabs

# Testing
pytest
pytest-asyncio
